feat(partner-ui): #342 — fold design/inventory work-order detail into unified /orders/:id
Design and inventory work-orders now open as fully-actionable unified orders
at /orders/:id instead of the retail order detail. Retail-only sections are
hidden for customer-less work-orders.

Backend:
- partners/orders/[id]: attach order↔execution reverse links
  (production_runs/inventory_orders) as the kind discriminator (metadata.kind
  retired). Done server-side via query.graph, not via DEFAULT_FIELDS — the
  admin order query-config rejects the custom links.
- partners/inventory-orders/[orderId]: expose unified_order_id via the
  inventory_orders.order.id forward link, for the legacy-detail redirect.

Partner UI:
- order-detail is kind-aware (use-order-kind helper); work-orders render
  WorkOrderStatusSection + design ProductionRunCard / inventory lines + actions.
- Shared components/work-orders/: ProductionRunCard extracted from the design
  page, InventoryOrderLines + InventoryOrderActionsSection extracted from the
  retired bespoke inventory detail; lib/work-status.ts shared label map.
- Inventory actions nested at /orders/:id/inventory/{start,complete,submit-payment};
  useInventoryActionTarget resolves the legacy id (router state → metadata.legacy_id)
  and invalidates the order detail on success so the work-status badge follows.
- /inventory-orders/:id retired → inventory-order-redirect to the unified order.
  /designs/:id kept as the design-management surface (deep-linked).

Tests: orders-unification-partner-detail.spec.ts 4/4 (own design/inventory/retail
detail + cross-partner 404); list-filter 4/4 regression; partner-ui vite build clean.
