--- apps/partner-ui/src/routes/inventory-orders/inventory-order-detail/index.ts
@@ -1,1 +1,0 @@
-export { InventoryOrderDetail as Component } from "./inventory-order-detail.tsx"
