+++ apps/partner-ui/src/routes/inventory-orders/inventory-order-redirect/index.ts
@@ -1,0 +1,1 @@
+export { InventoryOrderRedirect as Component } from "./inventory-order-redirect"
