+++ apps/partner-ui/src/routes/orders/order-detail/components/work-order-status-section/index.ts
@@ -1,0 +1,1 @@
+export { WorkOrderStatusSection } from "./work-order-status-section"
