+++ apps/partner-ui/src/routes/orders/order-detail/order-detail.tsx
@@ -1,8 +1,16 @@
+import { useQueryClient } from "@tanstack/react-query"
 import { useLoaderData, useParams } from "react-router-dom"
 
 import { TwoColumnPageSkeleton } from "../../../components/common/skeleton"
 import { TwoColumnPage } from "../../../components/layout/pages"
-import { useOrder, useOrderPreview } from "../../../hooks/api/orders"
+import { InventoryOrderActionsSection } from "../../../components/work-orders/inventory-order-actions-section"
+import { InventoryOrderLines } from "../../../components/work-orders/inventory-order-lines"
+import { ProductionRunCard } from "../../../components/work-orders/production-run-card"
+import { ordersQueryKeys, useOrder, useOrderPreview } from "../../../hooks/api/orders"
+import { usePartnerConsumptionLogs } from "../../../hooks/api/partner-consumption-logs"
+import { usePartnerDesign } from "../../../hooks/api/partner-designs"
+import { usePartnerInventoryOrder } from "../../../hooks/api/partner-inventory-orders"
+import { usePartnerProductionRun } from "../../../hooks/api/partner-production-runs"
 import { usePlugins } from "../../../hooks/api/plugins"
 import { useExtension } from "../../../providers/extension-provider"
 import { ActiveOrderClaimSection } from "./components/active-order-claim-section"
@@ -15,13 +23,16 @@ import { OrderFulfillmentSection } from "./components/order-fulfillment-section"
 import { OrderGeneralSection } from "./components/order-general-section"
 import { OrderPaymentSection } from "./components/order-payment-section"
 import { OrderSummarySection } from "./components/order-summary-section"
+import { WorkOrderStatusSection } from "./components/work-order-status-section"
 import { DEFAULT_FIELDS } from "./constants"
 import { orderLoader } from "./loader"
+import { useOrderKind } from "./use-order-kind"
 
 export const OrderDetail = () => {
   const initialData = useLoaderData() as Awaited<ReturnType<typeof orderLoader>>
 
   const { id } = useParams()
+  const queryClient = useQueryClient()
   const { getWidgets } = useExtension()
   const { plugins = [] } = usePlugins()
 
@@ -54,7 +65,32 @@ export const OrderDetail = () => {
     id!
   )
 
-  if (isLoading || !order || isPreviewLoading) {
+  // #342 — a unified order can be a design/inventory work-order. Derive the
+  // kind from the reverse execution links the detail route attaches, then fold
+  // in the work-specific surfaces (run card / inventory lines + actions) and
+  // hide the retail-only sections (edit/returns/exchanges/claims/payment/
+  // fulfillment/customer) — work-orders are customer-less.
+  const { kind, legacyId, isWorkOrder } = useOrderKind(order)
+
+  // Design work-order: legacy_id is the production_run id → resolve the run,
+  // then the design (for the Complete materials form, name, cost currency).
+  const { production_run } = usePartnerProductionRun(legacyId ?? "", {
+    enabled: kind === "design" && !!legacyId,
+  })
+  const designId = (production_run as any)?.design_id as string | undefined
+  const { design } = usePartnerDesign(designId ?? "", { enabled: !!designId })
+  const { logs: consumptionLogs = [], count: consumptionCount = 0 } =
+    usePartnerConsumptionLogs(designId ?? "", undefined, { enabled: !!designId })
+
+  // Inventory work-order: legacy_id is the inventory_order id.
+  const { inventoryOrder } = usePartnerInventoryOrder(legacyId ?? "", {
+    enabled: kind === "inventory" && !!legacyId,
+  })
+
+  const invalidateOrder = () =>
+    queryClient.invalidateQueries({ queryKey: ordersQueryKeys.detail(id!) })
+
+  if (isLoading || !order) {
     return (
       <TwoColumnPageSkeleton mainSections={4} sidebarSections={2} showJSON />
     )
@@ -62,6 +98,14 @@ export const OrderDetail = () => {
 
   if (isError) {
     throw error
+  }
+
+  // Retail order preview drives the active claim/exchange/return sections; work
+  // orders don't use it, so don't block their render on the preview fetch.
+  if (!isWorkOrder && isPreviewLoading) {
+    return (
+      <TwoColumnPageSkeleton mainSections={4} sidebarSections={2} showJSON />
+    )
   }
 
   return (
@@ -78,17 +122,46 @@ export const OrderDetail = () => {
       hasOutlet
     >
       <TwoColumnPage.Main>
-        <OrderActiveEditSection order={order} />
-        <ActiveOrderClaimSection orderPreview={orderPreview!} />
-        <ActiveOrderExchangeSection orderPreview={orderPreview!} />
-        <ActiveOrderReturnSection orderPreview={orderPreview!} />
-        <OrderGeneralSection order={order} />
-        <OrderSummarySection order={order} plugins={plugins} />
-        <OrderPaymentSection order={order} plugins={plugins} />
-        <OrderFulfillmentSection order={order} />
+        {!isWorkOrder && (
+          <>
+            <OrderActiveEditSection order={order} />
+            <ActiveOrderClaimSection orderPreview={orderPreview!} />
+            <ActiveOrderExchangeSection orderPreview={orderPreview!} />
+            <ActiveOrderReturnSection orderPreview={orderPreview!} />
+            <OrderGeneralSection order={order} />
+            <OrderSummarySection order={order} plugins={plugins} />
+            <OrderPaymentSection order={order} plugins={plugins} />
+            <OrderFulfillmentSection order={order} />
+          </>
+        )}
+        {isWorkOrder && (
+          <>
+            <WorkOrderStatusSection order={order} kind={kind} designId={designId} />
+            {kind === "design" && production_run && design && (
+              <ProductionRunCard
+                run={production_run}
+                design={design}
+                consumptionLogs={consumptionLogs}
+                consumptionCount={consumptionCount}
+                onActionSuccess={invalidateOrder}
+              />
+            )}
+            {kind === "inventory" && inventoryOrder && (
+              <InventoryOrderLines
+                orderLines={(inventoryOrder.order_lines ?? []) as Array<Record<string, any>>}
+              />
+            )}
+          </>
+        )}
       </TwoColumnPage.Main>
       <TwoColumnPage.Sidebar>
-        <OrderCustomerSection order={order} />
+        {kind === "inventory" && inventoryOrder && (
+          <InventoryOrderActionsSection
+            inventoryOrder={inventoryOrder}
+            linkPrefix="inventory/"
+          />
+        )}
+        {!isWorkOrder && <OrderCustomerSection order={order} />}
         <OrderActivitySection order={order} />
       </TwoColumnPage.Sidebar>
     </TwoColumnPage>
